Replace integration tests with real gradle builds
This commit uses assert_fs to run integration tests against actual (minimal)
gradle builds. This is much slower, but is more robust.

In addition, instead of searching for build.gradle[.kts] files and running from
there, it finds settings.gradle[.kts] and passes the --project-directory flag to gradle.
Specifying the build file is deprecated, and usually you want to run gradle from the root
of the project.